ctt: cut calibration peak memory from ~9.7 GB to ~1.5 GB
Loading a full calibration set exhausted the RAM of an 8 GB Pi and locked
the machine up: channels were stored as int64 holding 16-bit values, every
non-dark image kept a never-used full-res demosaic, and burst groups held
all frames in memory at once. Store channels as uint16, demosaic only for
CAC (the sole img.rgb consumer), average bursts one frame at a time and
close the rawpy handle promptly. The two consumer sites whose arithmetic
would wrap in uint16 (ALSC cell sums, patch black-level subtraction) now
promote to int64 locally, keeping the tuning output bit-identical.

+++ apps/ctt/ctt/algorithms/alsc.py
@@ -306,7 +306,8 @@ def get_grid(chan: np.ndarray, dx: int, dy: int, grid_size: tuple[int, int]) -> 
     row_edges = np.arange(grid_h) * dy
     col_edges = np.arange(grid_w) * dx
     # reduceat sums contiguous blocks between edges; the last block extends to the array boundary.
-    cell_sums = np.add.reduceat(np.add.reduceat(chan, row_edges, axis=0), col_edges, axis=1)
+    # Accumulate in int64: the channels are stored as uint16 and a cell sum would overflow.
+    cell_sums = np.add.reduceat(np.add.reduceat(chan, row_edges, axis=0, dtype=np.int64), col_edges, axis=1)
     row_sizes = np.diff(np.append(row_edges, h_total))
     col_sizes = np.diff(np.append(col_edges, w_total))
     cell_counts = row_sizes[:, np.newaxis] * col_sizes[np.newaxis, :]
